python-backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
requests
pytest
pytest-flask